sec(core): comprehensive type hardening and DoS mitigation
This commit applies multiple critical security fixes to the Value evaluator:

1. Integer Precision & Type Safety:
   - Split `Value::Decimal` and `Value::Int` to enforce precision bounds.
   - Removed `From<i64>` on `Value` to prevent silent coercion and precision loss.
   - `BinOp::Div` now explicitly promotes `Int` to `Decimal`.

2. Algorithmic Complexity / Hash Collision DoS:
   - Re-introduced `hash_field` utilizing the FNV-1a algorithm.
   - `RecordField` instances now pre-compute their `hash` on creation.
   - `Value::get_field` requires a pre-computed `field_hash` to filter
     comparisons in O(1), preventing worst-case string comparison DoS.
   - `ast::Expr::FieldAccess` caches the `field_hash` at parse time.

3. Budget Isolation & Hydration:
   - `budget_eq` is now isolated behind a local 10,000 instruction budget in
     the logic worker to prevent it from consuming the global evaluator budget.
   - `from_json_slice` now accepts an `is_trusted` flag. Internal local-disk
     storage hydration bypasses security budgets, while network hydration
     remains heavily guarded.

+++ src/render/window/event_loop.rs
@@ -1799,7 +1799,7 @@ fn dispatch_redraw_requested(
     // Expose scroll state to the logic store
     tab.store.set_runtime(
         "root_scroll_y",
-        crate::core::types::Value::Int(
+        crate::core::types::Value::Decimal(
             (tab.root_scroll_offset_y as f64 * crate::core::types::DECIMAL_SCALE as f64).round()
                 as i64,
         ),
